fixing typo and changing datatypes to float

+++ db/migrate/20200723224206_create_games.rb
@@ -3,8 +3,8 @@ class CreateGames < ActiveRecord::Migration[6.0]
     create_table :games do |t|
       t.integer :user_id
       t.integer :passage_id
-      t.integer :speed
-      t.integer :accuracy
+      t.float :speed
+      t.float :accuracy
 
       t.timestamps
     end
